src/main/java/ufo/codility/lessons/lesson_3/timeComplexity/PermMissingElem.java
/*******************************************************************************
 * Copyright 2018 Francesco Cina'
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 ******************************************************************************/
package ufo.codility.lessons.lesson_3.timeComplexity;

public class PermMissingElem {

    public int solution(int[] A) {
    	final long max = A.length + 1;
        long sum = max*(max+1)/2;
        for (final int element : A) {
        	sum -= element;
		}
        return (int) sum;
    }

}
